requirements.txt
celery==5.0.5
Django==3.1.7
django-allauth==0.44.0
django-filter==2.4.0
django-widget-tweaks==1.4.8
Pillow==8.1.0
redis
pylint
django-modeltranslation==0.16.2

